+++ toyerp/frontend/app/console/accounts/page.tsx
@@ -1,0 +1,2 @@
+import { AccountManagement } from "@/components/account-management"
+export default function Page() { return <AccountManagement /> }
