+++ toyerp/frontend/app/console/dashboard/page.tsx
@@ -1,0 +1,2 @@
+import { DashboardContent } from "@/components/dashboard-content"
+export default function Page() { return <DashboardContent /> }
