+++ toyerp/frontend/app/console/employees/page.tsx
@@ -1,0 +1,2 @@
+import { EmployeeManagement } from "@/components/employee-management"
+export default function Page() { return <EmployeeManagement /> }
