+++ toyerp/frontend/app/console/inventory/page.tsx
@@ -1,0 +1,2 @@
+import { InventoryManagement } from "@/components/inventory-management"
+export default function Page() { return <InventoryManagement /> }
