+++ toyerp/frontend/app/console/journal/page.tsx
@@ -1,0 +1,2 @@
+import { JournalEntry } from "@/components/journal-entry"
+export default function Page() { return <JournalEntry /> }
