+++ toyerp/frontend/app/console/reports/page.tsx
@@ -1,0 +1,2 @@
+import { Reports } from "@/components/reports"
+export default function Page() { return <Reports /> }
