+++ toyerp/frontend/app/console/user-approval/page.tsx
@@ -1,0 +1,2 @@
+import { UserApproval } from "@/components/user-approval"
+export default function Page() { return <UserApproval /> }
